test(translator): split responses chat request coverage
